audio_paralinguistic/requirements.txt
# ===========================================
# 语音副语言感知评估系统 - 依赖清单
# ===========================================
# 注意: 以下依赖需要根据实际情况安装
# 模型依赖体积较大，请按需安装
# ===========================================

# ============ 核心依赖 ============
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ============ 音频处理 ============
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# ============ 深度学习框架 ============
torch>=2.0.0
transformers>=4.35.0

# ============ ASR模型 ============
# FunASR (SenseVoiceSmall)
# pip install funasr modelscope

# Faster-Whisper
# pip install faster-whisper

# ============ 说话人识别 ============
# WeSpeaker (CAM++ / ECAPA-TDNN)
# pip install wespeaker

# ============ 情感识别 ============
# Emotion2Vec
# pip install emotion2vec

# ============ 聚类与机器学习 ============
scikit-learn>=1.3.0
hdbscan>=0.8.0

# ============ 降维可视化 ============
umap-learn>=0.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# ============ API与并发 ============
requests>=2.31.0
aiohttp>=3.8.0
tqdm>=4.65.0

# OpenAI API (用于评测)
openai>=1.0.0

# Google GenAI (用于Gemini)
# google-genai>=0.3.0

# ============ 配置管理 ============
pyyaml>=6.0
omegaconf>=2.3.0

# ============ 可选依赖 ============
# Jupyter支持
# jupyter>=1.0.0
# ipython>=8.0.0

# 数据集处理
# datasets>=2.14.0

# 分布式处理
# ray>=2.0.0
